refactor(session): split the 1154-line orchestrator_helpers.go by responsibility
The "helpers" file was a grab-bag of distinct Orchestrator responsibilities. Pure
same-package file reorganization (no behavior change) into cohesive siblings:

- orchestrator_helpers.go (1154 -> 92): resolveSession / buildSessionContext + shared types
- orchestrator_leases.go:      lease acquisition, pipeline start/stop, quality mapping
- orchestrator_recovery.go:    startup-recovery profile selection + fallback restart
- orchestrator_playlist.go:    playlist-readiness + the m3u8 segment/init parsing
- orchestrator_transitions.go: state transitions + execution loop + finalize

No behavior change. Build + go vet clean; session domain tests green; golangci-lint
--new-from-rev = 0; verify-config green.
